Move all files which uses SIMDataFormats to test
