+++ Calibration/HcalCalibAlgos/test/BuildFile.xml
@@ -1,4 +1,5 @@
 <use   name="Calibration/HcalCalibAlgos"/>
+<use   name="SimDataFormats/Track"/>
 <use   name="CondFormats/HcalObjects"/>
 <use   name="CondFormats/L1TObjects"/>
 <use   name="DQMServices/Core"/>
@@ -29,11 +30,12 @@
 <use   name="boost"/>
 <use   name="root"/>
 <use   name="DataFormats/CaloRecHit"/>
+<use   name="SimDataFormats/GeneratorProducts"/>
 <use   name="clhep"/>
 <use   name="hepmc"/>
 <use   name="CommonTools/UtilAlgos"/>
 <use   name="TrackPropagation/SteppingHelixPropagator"/>
 <use   name="TrackingTools/TrackAssociator"/>
-<library   file="*.cc" name="CalibrationHcalCalibAlgosPlugins">
+<library   file="*.cc" name="CalibrationHcalCalibAlgosTest">
   <flags   EDM_PLUGIN="1"/>
 </library>
