+++ Calibration/HcalCalibAlgos/test/DiJetAnalyzer.h
